feat: 예약있는 시간 및 테마 삭제 시 검증 추가

+++ src/main/java/roomescape/theme/service/ThemeService.java
@@ -2,6 +2,9 @@ package roomescape.theme.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.global.exception.BusinessException;
+import roomescape.global.exception.ErrorCode;
+import roomescape.reservation.repository.ReservationRepository;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.repository.ThemeRepository;
 import roomescape.time.domain.ReservationTime;
@@ -16,10 +19,12 @@ public class ThemeService {
 
     private final ThemeRepository themeRepository;
     private final ReservationTimeRepository reservationTimeRepository;
+    private final ReservationRepository reservationRepository;
 
-    public ThemeService(ThemeRepository themeRepository, ReservationTimeRepository reservationTimeRepository) {
+    public ThemeService(ThemeRepository themeRepository, ReservationTimeRepository reservationTimeRepository, ReservationRepository reservationRepository) {
         this.themeRepository = themeRepository;
         this.reservationTimeRepository = reservationTimeRepository;
+        this.reservationRepository = reservationRepository;
     }
 
     @Transactional
@@ -30,6 +35,9 @@ public class ThemeService {
 
     @Transactional
     public void removeTheme(Long id) {
+        if (reservationRepository.existsByThemeId(id)) {
+            throw new BusinessException(ErrorCode.THEME_IN_USE);
+        }
         themeRepository.remove(id);
     }
