+++ src/main/java/roomescape/time/service/ReservationTimeService.java
@@ -2,6 +2,9 @@ package roomescape.time.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.global.exception.BusinessException;
+import roomescape.global.exception.ErrorCode;
+import roomescape.reservation.repository.ReservationRepository;
 import roomescape.time.domain.ReservationTime;
 import roomescape.time.repository.ReservationTimeRepository;
 
@@ -13,9 +16,11 @@ import java.util.List;
 public class ReservationTimeService {
 
     private final ReservationTimeRepository reservationTimeRepository;
+    private final ReservationRepository reservationRepository;
 
-    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository) {
+    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository, ReservationRepository reservationRepository) {
         this.reservationTimeRepository = reservationTimeRepository;
+        this.reservationRepository = reservationRepository;
     }
 
     @Transactional
@@ -30,6 +35,9 @@ public class ReservationTimeService {
 
     @Transactional
     public void removeTime(Long id) {
+        if (reservationRepository.existsByTimeId(id)) {
+            throw new BusinessException(ErrorCode.RESERVATION_TIME_IN_USE);
+        }
         reservationTimeRepository.remove(id);
     }
 }
